+++ src/spelling/spellcheckdecorator.cpp
@@ -1,49 +1,67 @@
 /*
  * SPDX-FileCopyrightText: 2022 Megan Conkle <megan.conkle@kdemail.net>
+ * SPDX-FileCopyrightText: 2006 Jacob R Rideout <kde@jacobrideout.net>
+   SPDX-FileCopyrightText: 2006 Martin Sandsmark <martin.sandsmark@kde.org>
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
 #include <QAction>
 #include <QContextMenuEvent>
+#include <QList>
 #include <QMenu>
 #include <QStringList>
+#include <QStringRef>
+#include <QTextBlock>
+#include <QTextBoundaryFinder> 
 #include <QTextCharFormat>
 #include <QTextCursor>
 #include <QTextLayout>
 #include <QTimer>
 #include <QVector>
 
-#include "spellcheckdecorator.h"
+#include <Sonnet/BackgroundChecker>
+#include <Sonnet/Dialog>
+#include <Sonnet/GuessLanguage>
+#include <Sonnet/Settings>
+#include <Sonnet/Speller>
 
-#include "dictionary.h"
-#include "dictionarymanager.h"
-#include "spellchecker.h"
+#include "spellcheckdecorator.h"
+#include "spellcheckdialog.h"
 
 namespace ghostwriter
 {
-
 class SpellCheckDecoratorPrivate
 {
     Q_DECLARE_PUBLIC(SpellCheckDecorator)
 
 public:
+    struct Position {
+        int start, length;
+    };
+
+    /**
+     * This structure abstracts the positions of breaks in the test. As per the
+     * unicode annex, both the start and end of the text are returned.
+     */
+    typedef QList<Position> Positions;
+
     SpellCheckDecoratorPrivate(SpellCheckDecorator *decorator)
-    : q_ptr(decorator),
-      spellCheckEnabled(true),
-      dictionary(DictionaryManager::instance()->requestDictionary())
+    : q_ptr(decorator)
     {
         ;
     }
     ~SpellCheckDecoratorPrivate() { }
 
+    static Sonnet::Settings *settings;
+
     SpellCheckDecorator *q_ptr;
     QPlainTextEdit *editor;
-    bool spellCheckEnabled;
-    Dictionary *dictionary;
+    Sonnet::Speller *speller;
     QColor errorColor;
+    SpellCheckDialog *spellCheckDialog;
 
-    QMenu * createContextMenu(const QTextCursor &cursorForWord) const;
+    QMenu * createContextMenu();
 
     QMenu * createSpellingMenu(
         const QString &misspelledWord,
@@ -55,8 +73,12 @@ public:
     void spellCheckBlock(QTextBlock &block) const;
     void clearSpellCheckFormatting(QTextBlock &block) const;
     void resetLiveSpellChecking() const;
+    Positions wordBreaks(const QString &text) const;
+    Positions sentenceBreaks(const QString &text) const;
 
 };
+
+Sonnet::Settings *SpellCheckDecoratorPrivate::settings = nullptr;
 
 SpellCheckDecorator::SpellCheckDecorator(QPlainTextEdit *editor)
 : QObject(editor),
@@ -64,7 +86,10 @@ SpellCheckDecorator::SpellCheckDecorator(QPlainTextEdit *editor)
 {
     Q_D(SpellCheckDecorator);
 
-    d->spellCheckEnabled = true;
+    if (nullptr == d->settings) {
+        d->settings = new Sonnet::Settings();
+    }
+    
     d->editor = editor;
 
     Q_ASSERT(nullptr != d->editor);
@@ -79,6 +104,8 @@ SpellCheckDecorator::SpellCheckDecorator(QPlainTextEdit *editor)
             d->onContentsChanged(position, charsAdded, charsRemoved);
         }
     );
+
+    d->speller = new Sonnet::Speller();
 }
 
 SpellCheckDecorator::~SpellCheckDecorator()
@@ -86,35 +113,11 @@ SpellCheckDecorator::~SpellCheckDecorator()
     ;
 }
 
-bool SpellCheckDecorator::liveSpellCheckEnabled() const
-{
-    Q_D(const SpellCheckDecorator);
-
-    return d->spellCheckEnabled;
-}
-
 QColor SpellCheckDecorator::errorColor() const
 {
     Q_D(const SpellCheckDecorator);
 
     return d->errorColor;
-}
-
-void SpellCheckDecorator::setLiveSpellCheckEnabled(bool enabled)
-{
-    Q_D(SpellCheckDecorator);
-
-    if (d->spellCheckEnabled != enabled) {
-        d->spellCheckEnabled = enabled;
-        d->resetLiveSpellChecking();
-    }
-}
-
-void SpellCheckDecorator::setDictionary(const QString &language)
-{
-    Q_D(SpellCheckDecorator);
-
-    d->dictionary = DictionaryManager::instance()->requestDictionary(language);
 }
 
 void SpellCheckDecorator::setErrorColor(const QColor &color)
@@ -125,22 +128,24 @@ void SpellCheckDecorator::setErrorColor(const QColor &color)
     d->resetLiveSpellChecking();
 }
 
-void SpellCheckDecorator::runSpellCheck()
+void SpellCheckDecorator::settingsChanged()
 {
     Q_D(SpellCheckDecorator);
 
-    SpellChecker::checkDocument(d->editor, d->dictionary);
-
-    if (d->spellCheckEnabled) {
-        d->resetLiveSpellChecking();
+    if (d->settings) {
+        delete d->settings;
+        d->settings = new Sonnet::Settings(this);
     }
+
+    d->speller->setLanguage(d->settings->defaultLanguage());
+    d->resetLiveSpellChecking();
 }
 
 void SpellCheckDecorator::startLiveSpellCheck()
 {
     Q_D(SpellCheckDecorator);
 
-    if (!d->spellCheckEnabled) {
+    if (!d->settings->checkerEnabledByDefault()) {
         return;
     }
 
@@ -158,7 +163,7 @@ bool SpellCheckDecorator::eventFilter(QObject *watched, QEvent *event)
     Q_UNUSED(watched);
 
     if ((event->type() != QEvent::ContextMenu)
-            || !d->spellCheckEnabled
+            || !d->settings->checkerEnabledByDefault()
             || d->editor->isReadOnly()) {
         return false;
     }
@@ -180,7 +185,7 @@ bool SpellCheckDecorator::eventFilter(QObject *watched, QEvent *event)
         cursorForWord = d->editor->cursorForPosition(contextEvent->pos());
     }
 
-    QMenu *popupMenu = d->createContextMenu(cursorForWord);
+    QMenu *popupMenu = d->createContextMenu();
     QString misspelledWord = d->getMisspelledWordAtCursor(cursorForWord);
 
     // If the selected word is spelled correctly, use the default processing for
@@ -215,10 +220,9 @@ bool SpellCheckDecorator::eventFilter(QObject *watched, QEvent *event)
     return true;
 }
 
-QMenu * SpellCheckDecoratorPrivate::createContextMenu(
-    const QTextCursor &cursorForWord) const
+QMenu * SpellCheckDecoratorPrivate::createContextMenu()
 {
-    Q_Q(const SpellCheckDecorator);
+    Q_Q(SpellCheckDecorator);
 
     // Add spell check action to the standard context menu that comes with
     // the editor.
@@ -230,9 +234,9 @@ QMenu * SpellCheckDecoratorPrivate::createContextMenu(
     q->connect(checkSpellingAction,
         &QAction::triggered,
         q,
-        [this, cursorForWord]() {
-            this->editor->setTextCursor(cursorForWord);
-            SpellChecker::checkDocument(this->editor, this->dictionary);
+        [this, q]() {
+            spellCheckDialog = new SpellCheckDialog(this->editor);
+            spellCheckDialog->show();
         }
     );
 
@@ -245,20 +249,40 @@ QMenu * SpellCheckDecoratorPrivate::createSpellingMenu(
     const QString &misspelledWord,
     const QTextCursor &cursorForWord) const
 {
-    Q_Q(const SpellCheckDecorator);
 
+    Q_Q(const SpellCheckDecorator);
     QMenu *spellingMenu = new QMenu(SpellCheckDecorator::tr("Spelling"));
-    QStringList suggestions = this->dictionary->suggestions(misspelledWord);
+
+    if (this->settings->autodetectLanguage())  {
+        QString text = cursorForWord.block().text();
+        int pos = cursorForWord.position() - cursorForWord.block().position();
+
+        for (auto sentenceBreak : sentenceBreaks(text)) {
+            if (pos < (sentenceBreak.start + sentenceBreak.length)) {
+                QString sentence = text.mid(
+                    sentenceBreak.start,
+                    sentenceBreak.length);
+                QString language = Sonnet::GuessLanguage().identify(sentence);
+
+                if (!language.isNull()) {
+                    this->speller->setLanguage(language);
+                }
+
+                break;
+            }
+        }
+    }
+
+    QStringList suggestions = this->speller->suggest(misspelledWord);
 
     QAction *addWordToDictionaryAction =
         new QAction(SpellCheckDecorator::tr("Add word to dictionary"), spellingMenu);
 
     q->connect(addWordToDictionaryAction,
         &QAction::triggered,
-        q,
         [this, cursorForWord, misspelledWord]() {
             this->editor->setTextCursor(cursorForWord);
-            this->dictionary->addToPersonal(misspelledWord);
+            this->speller->addToPersonal(misspelledWord);
 
             resetLiveSpellChecking();
         }
@@ -350,7 +374,7 @@ void SpellCheckDecoratorPrivate::onContentsChanged(
     int charsAdded,
     int charsRemoved)
 {
-    if (!this->spellCheckEnabled) {
+    if (!this->settings->checkerEnabledByDefault()) {
         return;
     }
 
@@ -380,27 +404,41 @@ void SpellCheckDecoratorPrivate::onContentsChanged(
 
 void SpellCheckDecoratorPrivate::spellCheckBlock(QTextBlock &block) const
 {
-    QStringRef misspelledWord = dictionary->check(block.text(), 0);
+    QString text = block.text();
 
-    while (!misspelledWord.isNull()) {
-        int startIndex = misspelledWord.position();
-        int length = misspelledWord.length();
+    for (auto sentenceSegment : sentenceBreaks(text)) {
+        QString sentence =
+            text.mid(sentenceSegment.start, sentenceSegment.length);
 
-        QTextCharFormat spellingErrorFormat;
-        spellingErrorFormat.setUnderlineColor(this->errorColor);
-        spellingErrorFormat.setUnderlineStyle(QTextCharFormat::SpellCheckUnderline);
+        if (this->settings->autodetectLanguage()) {
+            QString language = Sonnet::GuessLanguage().identify(sentence);
 
-        QTextLayout::FormatRange range;
-        range.start = startIndex;
-        range.length = length;
-        range.format = spellingErrorFormat;
+            if (!language.isNull()) {
+                this->speller->setLanguage(language);
+            } else {
+                this->speller->setLanguage(this->settings->defaultLanguage());
+            }
+        }
 
-        auto formats = block.layout()->formats();
-        formats.append(range);
-        block.layout()->setFormats(formats);
+        for (auto wordSegment : wordBreaks(sentence)) {
+            int wordStart = sentenceSegment.start + wordSegment.start;
+            QString word = text.mid(wordStart, wordSegment.length);
 
-        startIndex += length;
-        misspelledWord = dictionary->check(block.text(), startIndex);
+            if (speller->isMisspelled(word)) {
+                QTextCharFormat spellingErrorFormat;
+                spellingErrorFormat.setUnderlineColor(this->errorColor);
+                spellingErrorFormat.setUnderlineStyle(QTextCharFormat::SpellCheckUnderline);
+
+                QTextLayout::FormatRange range;
+                range.start = wordStart;
+                range.length = wordSegment.length;
+                range.format = spellingErrorFormat;
+
+                auto formats = block.layout()->formats();
+                formats.append(range);
+                block.layout()->setFormats(formats);
+            }
+        }
     }
 }
 
@@ -424,12 +462,80 @@ void SpellCheckDecoratorPrivate::resetLiveSpellChecking() const
     while (block.isValid()) {
         clearSpellCheckFormatting(block);
 
-        if (spellCheckEnabled) {
+        if (this->settings->checkerEnabledByDefault()) {
             spellCheckBlock(block);
         }
 
         block = block.next();
     }
+}
+
+// Code is lifted from KDE Frameworks' Sonnet library, because we know it
+// just works.  :)
+SpellCheckDecoratorPrivate::Positions SpellCheckDecoratorPrivate::wordBreaks(const QString &text) const
+{
+    Positions breaks;
+
+    if (text.isEmpty()) {
+        return breaks;
+    }
+
+    QTextBoundaryFinder boundaryFinder(QTextBoundaryFinder::Word, text);
+
+    while (boundaryFinder.position() < text.length()) {
+        if (!(boundaryFinder.boundaryReasons().testFlag(QTextBoundaryFinder::StartOfItem))) {
+            if (boundaryFinder.toNextBoundary() == -1) {
+                break;
+            }
+            continue;
+        }
+
+        Position pos;
+        pos.start = boundaryFinder.position();
+        int end = boundaryFinder.toNextBoundary();
+        if (end == -1) {
+            break;
+        }
+        pos.length = end - pos.start;
+        if (pos.length < 1) {
+            continue;
+        }
+        breaks.append(pos);
+
+        if (boundaryFinder.toNextBoundary() == -1) {
+            break;
+        }
+    }
+    return breaks;
+}
+
+// Code is lifted from KDE Frameworks' Sonnet library, because we know it
+// just works.  :)
+SpellCheckDecoratorPrivate::Positions
+SpellCheckDecoratorPrivate::sentenceBreaks(const QString &text) const
+{
+    Positions breaks;
+
+    if (text.isEmpty()) {
+        return breaks;
+    }
+
+    QTextBoundaryFinder boundaryFinder(QTextBoundaryFinder::Sentence, text);
+
+    while (boundaryFinder.position() < text.length()) {
+        Position pos;
+        pos.start = boundaryFinder.position();
+        int end = boundaryFinder.toNextBoundary();
+        if (end == -1) {
+            break;
+        }
+        pos.length = end - pos.start;
+        if (pos.length < 1) {
+            continue;
+        }
+        breaks.append(pos);
+    }
+    return breaks;
 }
 
 } // namespace ghostwriter